README.md update sequence number assigning refactoring fix: bookname was not assigned for SECOND direction messages dependency: kafka client update
