+++ .github/workflows/ci-unwelcome-words.yml
@@ -7,17 +7,17 @@ jobs:
   test:
     runs-on: ubuntu-20.04
     steps:
-    - uses: actions/checkout@v2
-      with:
-        ref: ${{ github.sha }}
-    - name: Checkout tool
-      uses: actions/checkout@v2
-      with:
-        repository: exactpro-th2/ci-github-action
-        ref: master
-        token: ${{ secrets.PAT_CI_ACTION }}
-        path: ci-github-action
-    - name: Run CI action
-      uses: ./ci-github-action
-      with:
-        ref: ${{ github.sha }}
+      - uses: actions/checkout@v2
+        with:
+          ref: ${{ github.sha }}
+      - name: Checkout tool
+        uses: actions/checkout@v2
+        with:
+          repository: exactpro-th2/ci-github-action
+          ref: master
+          token: ${{ secrets.PAT_CI_ACTION }}
+          path: ci-github-action
+      - name: Run CI action
+        uses: ./ci-github-action
+        with:
+          ref: ${{ github.sha }}
