+++ build.gradle
@@ -46,11 +46,9 @@ jar {
 
 dependencies {
     api platform('com.exactpro.th2:bom:4.1.0')
-
     implementation 'com.exactpro.th2:common:5.1.0-dev-version-5-4085018593-8adee33-SNAPSHOT'
 
-    implementation 'org.apache.kafka:kafka_2.13:3.3.1'
-    implementation 'io.github.microutils:kotlin-logging:3.0.0'
+    implementation 'org.apache.kafka:kafka-clients:3.4.0'
 
     testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
     testImplementation 'org.assertj:assertj-core:3.24.2'
